Add a hook to support additional UI on the comment issue bar on review page.
This hook is called CommentIssueBarActionHook. It allows an extension to render
additional UI for a comment in the comment issue bar. The hook makes itself and
the CommentIssueManager instance available to the extension. The extension can
use the CommentIssueBar instance to manage the UI elements and the
CommentIssueManager instance to set or get the extra_data for the comment.

CommentIssueManager exposes two new APIs getCommentExtraData and
setCommentExtraData which can be used by the extension to manage custom fields.

NOTE: This hook has been published for inclusion in reviewboard mainline via
review request https://reviews.reviewboard.org/r/7841/